Atualização do github
Pasta com os códigos antigos foi renomeada como "original"
Nova pasta para os códigos atualizados.
Atualização do código main para o rastreador.
