--- application_gui/data2.csv
@@ -1,13 +1,0 @@
--25.52;-24.33;-22.17;-21.56;-21.36;-21.2;-21.09
--21.61;-20.32;-19.14;-19.1;-22.06;-21.02;-21.77
--18.03;-23.21;-35.62;-38.97;-40.59;-41.11;-41.8
--16.52;-24.63;-44.29;-51.25;-53.97;-55.51;-55.43
--16.4;-24.84;-44.76;-52.32;-55.68;-56.03;-57.32
--16.15;-25.32;-45.32;-51.43;-56.72;-57.23;-56.23
--16.37;-25.5;-44.56;-53.77;-55.77;-57.04;-58.17
--16.63;-26.15;-43.92;-54.48;-52.98;-58.02;-57.16
--17.12;-25.15;-43.64;-53.98;-54.87;-56.01;-57.49
--17.47;-23.95;-44.93;-54.96;-57.85;-58.01;-56.13
--18.05;-27.16;-45.35;-53.97;-57.38;-57.64;-55.77
--22.65;-32.46;-45.88;-52.68;-55.48;-56.16;-55.43
--29.19;-34.59;-36.03;-42.55;-44.79;-46.61;-46.44
