--- application_gui/plot_magnetico_teste/eixoY.csv
@@ -1,13 +1,0 @@
--58.85;-57.96;-56.84;-55.82;-54.92;-53.81;-53.24;-54.18;-66.41;-45.02;-35.98;-39.56;-51.49
--57.35;-55.67;-54.54;-53.6;-52.39;-51.18;-50.04;-49.99;-56.31;-44.78;-34.63;-37.65;-48.92
--55.8;-54.07;-52.81;-51.68;-50.46;-49.27;-48.02;-47.72;-52.72;-44.39;-33.66;-36.17;-46.26
--54.26;-52.45;-51.23;-50.21;-49.18;-47.95;-46.83;-46.7;-51.96;-43.56;-33.14;-35.48;-44.54
--53.01;-51.01;-49.89;-49.13;-48.29;-47.19;-46.26;-46.52;-52.96;-42.57;-32.89;-35.31;-43.63
--51.58;-49.8;-48.81;-48.39;-47.7;-46.91;-46.29;-46.81;-55.01;-41.97;-32.84;-35.29;-43.26
--50.34;-48.51;-47.8;-47.65;-47.43;-47.0;-46.83;-47.87;-58.62;-41.34;-32.86;-35.38;-43.15
--49.06;-47.2;-46.91;-47.31;-47.67;-47.75;-47.9;-49.49;-62.99;-40.68;-32.97;-35.68;-43.24
--47.78;-45.92;-46.29;-47.63;-48.73;-49.24;-49.99;-52.7;-58.47;-40.02;-33.16;-36.19;-43.68
--47.31;-45.53;-46.78;-49.22;-51.1;-52.47;-53.98;-58.75;-54.78;-40.42;-34.14;-37.29;-44.88
--50.19;-48.53;-49.15;-51.13;-52.7;-54.02;-55.26;-59.57;-57.58;-43.47;-37.54;-40.63;-47.67
--57.01;-53.55;-51.66;-51.35;-51.58;-51.94;-52.7;-54.88;-61.09;-51.68;-45.53;-47.17;-52.81
--59.01;-57.71;-56.8;-55.7;-55.18;-55.45;-56.21;-58.21;-63.8;-61.53;-55.66;-55.7;-60.0
