+++ application_gui_original/data.csv
@@ -1,0 +1,13 @@
+-25.52;-24.33;-22.17;-21.56;-21.36;-21.2;-21.09;-21.39;-21.31;-21.91;-21.79;-23.4;-26.32
+-21.61;-20.32;-19.14;-19.1;-22.06;-21.02;-21.77;-22.01;-22.77;-24.04;-27.46;-17.59;-22.15
+-18.03;-23.21;-35.62;-38.97;-40.59;-41.11;-41.8;-41.9;-42.38;-42.18;-39.95;-20.44;-23.01
+-16.52;-24.63;-44.29;-51.25;-53.97;-55.51;-55.43;-55.87;-55.8;-54.39;-50.39;-33.33;-28.17
+-16.4;-24.84;-44.76;-52.32;-55.68;-56.03;-57.32;-57.16;-58.66;-56.43;-55.84;-44.21;-35.61
+-16.15;-25.32;-45.32;-51.43;-56.72;-57.23;-56.23;-56.1;-59.02;-57.61;-54.21;-48.11;-40.72
+-16.37;-25.5;-44.56;-53.77;-55.77;-57.04;-58.17;-58.47;-58.17;-58.29;-57.22;-50.73;-43.11
+-16.63;-26.15;-43.92;-54.48;-52.98;-58.02;-57.16;-57.35;-54.44;-57.46;-56.97;-52.75;-44.02
+-17.12;-25.15;-43.64;-53.98;-54.87;-56.01;-57.49;-57.68;-57.68;-57.92;-56.03;-51.42;-45.79
+-17.47;-23.95;-44.93;-54.96;-57.85;-58.01;-56.13;-57.82;-59.49;-58.64;-53.66;-53.34;-45.74
+-18.05;-27.16;-45.35;-53.97;-57.38;-57.64;-55.77;-56.64;-56.64;-58.06;-55.17;-52.28;-45.54
+-22.65;-32.46;-45.88;-52.68;-55.48;-56.16;-55.43;-55.18;-56.39;-55.07;-54.78;-51.64;-44.95
+-29.19;-34.59;-36.03;-42.55;-44.79;-46.61;-46.44;-47.54;-49.2;-48.02;-47.88;-47.93;-43.76
