+++ application_gui_original/plot_magnetico_teste/eixoX.csv
@@ -1,0 +1,13 @@
+-62.56;-60.88;-58.5;-56.33;-54.17;-51.91;-49.67;-47.52;-45.65;-44.65;-46.65;-54.95;-60.8
+-62.16;-59.62;-57.28;-54.78;-52.65;-50.54;-48.51;-46.62;-45.06;-44.19;-45.35;-50.31;-57.18
+-61.59;-59.09;-56.51;-54.18;-52.06;-50.28;-48.73;-47.27;-46.16;-45.68;-46.81;-50.8;-55.36
+-61.5;-58.61;-56.14;-53.98;-52.34;-50.93;-49.8;-48.99;-48.56;-48.99;-50.7;-54.26;-56.73
+-61.1;-58.24;-55.75;-53.78;-52.48;-51.6;-50.91;-50.91;-51.58;-52.83;-55.72;-58.24;-59.13
+-60.86;-57.84;-55.27;-53.46;-52.21;-51.73;-51.73;-52.47;-53.98;-56.5;-59.55;-59.48;-61.25
+-60.95;-57.26;-54.54;-52.81;-51.91;-51.79;-52.16;-53.41;-55.34;-58.22;-60.61;-59.03;-62.53
+-62.42;-57.58;-54.35;-52.81;-52.42;-52.81;-53.78;-55.24;-56.48;-57.58;-57.7;-56.8;-62.36
+-66.66;-62.17;-58.36;-56.85;-56.96;-56.53;-54.92;-52.57;-50.7;-49.89;-50.59;-52.19;-58.36
+-56.21;-53.29;-50.18;-47.5;-45.7;-43.99;-42.56;-41.41;-40.54;-40.68;-42.4;-46.31;-53.53
+-49.43;-44.36;-40.89;-39.11;-37.97;-37.04;-36.37;-35.76;-35.31;-35.68;-37.87;-43.15;-51.15
+-50.93;-47.5;-44.7;-42.56;-41.16;-40.16;-39.33;-38.67;-38.37;-38.67;-40.81;-45.58;-52.59
+-57.06;-55.7;-53.49;-50.99;-49.04;-47.52;-46.26;-45.47;-45.13;-45.58;-47.36;-51.04;-56.61
