+++ application_gui_original/plot_magnetico_teste/mag_norm.png
