+++ application_gui_original/plot_magnetico_teste/normalizada.csv
@@ -1,0 +1,13 @@
+-57.3100279673007;-56.16877832814975;-54.58086566095089;-53.0572180611916;-51.51853007029796;-49.74661409826663;-48.08775775666923;-46.67146990978013;-45.61369488714579;-41.82076093570451;-35.62288648315081;-39.43623963031334;-51.008614784993476
+-56.11027706896674;-54.20029928989699;-52.68710546230146;-51.13974622116511;-49.507754649449474;-47.83792145984095;-46.19766909606625;-44.975708300369526;-44.74595809658071;-41.46468861517497;-34.2768125468315;-37.420769514395495;-48.31574362163873
+-54.78378993312299;-52.88148618162302;-51.26704015374429;-49.74224079643751;-48.17643029041761;-46.73540513305846;-45.35020700800014;-44.4788742303082;-45.29355311014064;-41.97697844423396;-33.45465847542752;-36.02296796477522;-45.75609849768468
+-53.508915196199325;-51.50842422613057;-50.01489591326728;-48.687853744949166;-47.468416292176585;-46.17896391711668;-45.055612496739926;-44.685479558760726;-46.92514161954083;-42.466179602923184;-33.064490194648904;-35.4228620987653;-44.28532384865048
+-52.383282704778736;-50.25732550734724;-48.888301259121704;-47.84999086170623;-46.887856218881;-45.847454155815385;-44.97999086170623;-45.172126828249915;-49.20511606630309;-42.17907671560268;-32.86742358386694;-35.287936182663366;-43.509292307505575
+-51.09545741324597;-49.16653022945421;-47.92529517380278;-47.213404549959904;-46.38382188497201;-45.672758212301865;-45.19840408686852;-45.76634478497286;-51.4542361836026;-41.81960177907782;-32.83074616955875;-35.27348197389156;-43.19155252330311
+-49.97811992382418;-47.96635129770994;-46.965561093812724;-46.49459547186099;-46.105958472142916;-45.755301883177864;-45.71371473720994;-46.800430859857194;-53.667139507653545;-41.25182045088265;-32.85271515998714;-35.36129968337104;-43.100192863463
+-48.86413613812429;-46.8192770086251;-46.19006940096732;-46.231667586178425;-46.41529977578308;-46.57102923185158;-46.90241338210902;-48.46541422244976;-55.6044670975552;-40.592221532242206;-32.95541002296774;-35.64657188914791;-43.18713861216573
+-47.72415439005441;-45.81821468943506;-46.02839946406193;-47.13908385837808;-48.12183520327835;-48.496817595721225;-48.77976854489183;-49.62421364041835;-50.02888559066357;-39.59409284745476;-33.082216171509415;-36.08225774488042;-43.534623170413326
+-46.78373398108661;-44.85745235702937;-45.14514161954083;-45.26510118880785;-44.59948221944762;-43.41369739349428;-42.257603501604926;-41.33060221734206;-40.37940579941584;-37.537754649449475;-33.53574362163872;-36.77726232705777;-44.32446489969762
+-46.783096548689144;-42.95232858092735;-40.28574362163873;-38.84546069538293;-37.82626004935389;-36.95380787706385;-36.31428201098759;-35.74197469034761;-35.28432562028973;-35.011743588071226;-34.69156640316476;-38.69942889153707;-46.06007932998865
+-49.97271507003073;-46.53676398662493;-43.903223406038734;-42.02103868679637;-40.782620162561216;-39.88090327252384;-39.134576627800264;-38.567283932123225;-38.34684599638438;-38.45809272518132;-39.5477527698602;-43.2923382632579;-49.68830712830211
+-54.91616248243467;-53.579439873463244;-51.82670735799719;-49.72522842230174;-48.09451871830801;-46.87143629701643;-45.8415038320151;-45.244846854413446;-45.0714062451505;-45.471025671506524;-46.76091843395922;-49.76254136977191;-54.97200214276592
