+++ application_gui_original/plot_magnetico_teste/teste_mag_X_08-12-2021_20-38.csv
@@ -1,0 +1,13 @@
+-43.75;-42.04;-39.63;-37.63;-36.2;-34.98;-33.92;-33.32;-33.32;-33.56;-33.56;-33.92;-35.0
+-43.6;-41.4;-39.6;-38.49;-37.51;-37.24;-36.8;-36.78;-37.22;-37.97;-38.68;-39.46;-37.87
+-45.84;-46.34;-44.49;-43.33;-42.89;-42.59;-42.51;-42.94;-44.04;-45.89;-47.82;-48.87;-48.7
+-50.73;-49.54;-49.19;-49.56;-50.88;-52.39;-54.84;-57.02;-53.34;-49.77;-48.79;-49.59;-52.29
+-53.65;-57.41;-59.38;-60.78;-52.76;-48.27;-45.63;-43.54;-41.7;-40.56;-40.09;-40.37;-41.43
+-56.34;-51.74;-50.41;-43.94;-42.51;-40.24;-38.41;-37.17;-36.09;-35.63;-35.63;-36.4;-37.27
+-56.54;-50.31;-45.84;-42.22;-39.01;-36.52;-34.79;-33.56;-32.44;-31.9;-31.8;-31.99;-32.71
+-46.65;-42.59;-39.28;-36.24;-33.72;-31.9;-30.38;-29.38;-28.71;-28.45;-28.58;-29.45;-30.97
+-46.01;-41.43;-37.44;-34.21;-31.6;-30.22;-28.58;-27.23;-26.14;-25.56;-25.47;-25.85;-26.76
+-40.37;-35.86;-32.29;-29.75;-27.94;-26.47;-25.31;-24.44;-23.96;-23.96;-24.44;-25.58;-26.76
+-43.82;-40.21;-38.18;-36.14;-33.94;-28.86;-27.46;-26.37;-25.58;-25.22;-25.31;-25.97;-26.96
+-43.72;-39.97;-36.86;-34.04;-32.24;-30.84;-29.55;-28.66;-28.4;-28.53;-28.91;-29.78;-28.53
+-45.99;-44.11;-41.33;-38.81;-36.99;-35.56;-35.93;-33.14;-32.6;-32.39;-32.6;-33.14;-34.21
