+++ application_gui_original/plot_magnetico_teste/teste_mag_Y_08-12-2021_20-38.csv
@@ -1,0 +1,13 @@
+-35.34;-32.6;-28.82;-27.23;-27.46;-28.58;-28.82;-28.89;-30.69;-37.24;-33.82;-29.36;-28.42
+-31.42;-27.94;-26.47;-26.58;-27.58;-27.82;-27.82;-28.71;-31.99;-38.21;-30.22;-28.04;-27.8
+-34.09;-30.3;-27.23;-25.99;-26.47;-27.12;-27.23;-27.34;-28.96;-35.14;-34.55;-29.03;-27.65
+-32.27;-28.69;-26.14;-25.94;-26.55;-26.79;-26.89;-27.58;-30.38;-42.27;-31.8;-28.27;-27.53
+-33.58;-29.83;-27.7;-25.77;-25.68;-26.27;-26.68;-26.76;-27.56;-30.61;-38.49;-29.96;-27.92
+-31.12;-27.56;-25.85;-25.75;-26.34;-26.66;-26.76;-27.46;-30.38;-40.49;-30.55;-28.32;-27.82
+-33.7;-30.78;-27.43;-26.04;-26.04;-26.66;-26.87;-26.98;-27.92;-31.31;-39.14;-30.78;-28.71
+-32.04;-28.66;-26.58;-26.47;-27.12;-27.46;-27.46;-27.82;-30.3;-39.06;-32.71;-29.9;-28.89
+-34.86;-32.34;-29.05;-27.41;-27.58;-28.19;-28.45;-28.45;-29.03;-31.99;-43.29;-33.72;-31.33
+-35.1;-31.68;-29.36;-29.15;-29.58;-29.96;-29.96;-30.35;-32.36;-39.25;-37.92;-34.09;-31.8
+-37.8;-36.54;-33.82;-32.39;-32.19;-32.5;-32.6;-32.71;-33.48;-35.93;-43.76;-41.43;-37.63
+-40.21;-37.99;-36.54;-35.9;-35.9;-35.96;-36.27;-37.22;-39.46;-46.34;-46.21;-41.75;-38.21
+-42.57;-43.24;-41.53;-40.44;-39.94;-39.94;-40.07;-40.63;-42.43;-45.93;-54.9;-50.36;-45.77
