+++ application_gui_original/scpii_test/README.md
@@ -1,0 +1,6 @@
+https://rsinstrument.readthedocs.io/en/latest/index.html
+
+https://www.rohde-schwarz.com/br/manual/hms/
+
+https://pyvisa.readthedocs.io/en/latest/introduction/communication.html
+
